Annotation update

+++ annotations/nocedit.exe/pseudocode/include/types/structs/SRenderVertex.h
@@ -10,9 +10,9 @@ typedef struct SRenderVertex {
     SProjectedVertex projected_vertex; // 0x0
     int u; // 0x18
     int v; // 0x1c
-    float light; // 0x20
+    int light; // 0x20
     int color; // 0x24
-    float fog; // 0x28
+    int fog; // 0x28
     float w_recip; // 0x2c
 } SRenderVertex;
