+++ annotations/nocedit.exe/pseudocode/src/core/dskybox.cpp/generateSkyDomeVertex_FUN_004900b0.cpp
@@ -2,13 +2,11 @@
 // Address: 004900b0
 // Address Range: [[004900b0, 004901ea]]
 // Convention: __cdecl
-// Signature: void __cdecl core_dskybox_cpp_generateSkyDomeVertex_FUN_004900b0 (float *light_color,int u_coord,int v_coord,float *uv_offset)
+// Signature: void __cdecl core_dskybox_cpp_generateSkyDomeVertex_FUN_004900b0(float *light_color,int u_coord,int v_coord,float *uv_offset)
 
 #include "nocturne.h"
 
-void __cdecl
-core_dskybox_cpp_generateSkyDomeVertex_FUN_004900b0
-          (float *light_color,int u_coord,int v_coord,float *uv_offset)
+void __cdecl core_dskybox_cpp_generateSkyDomeVertex_FUN_004900b0(float *light_color,int u_coord,int v_coord,float *uv_offset)
 
 {
   float fVar1;
@@ -46,9 +44,9 @@ core_dskybox_cpp_generateSkyDomeVertex_FUN_004900b0
   wincore_windll_cpp_transformAndProjectPoint_FUN_005b575c
             (&pCVar2->vertex_buffer_ptr[(int)uv_offset].projected_vertex,input);
   fVar1 = *light_color;
-  g_RenderVertexBuffer[(int)uv_offset].light = fVar1;
+  g_RenderVertexBuffer[(int)uv_offset].light = (int)fVar1;
   g_RenderVertexBuffer[(int)uv_offset].color = (int)fVar1;
-  g_RenderVertexBuffer[(int)uv_offset].fog = fVar1;
+  g_RenderVertexBuffer[(int)uv_offset].fog = (int)fVar1;
   g_RenderVertexBuffer[(int)uv_offset].w_recip = 0.0;
   return;
 }
